Refactor CSS into better components
