+++ client/component/error/style.scss
@@ -1,0 +1,24 @@
+@import "../../global";
+
+.notice-error {
+	margin-top: 3em;
+
+	.closer {
+		float: right;
+		padding-top: 5px;
+		font-size: 18px;
+		cursor: pointer;
+		color: #333;
+	}
+
+	textarea {
+		font-family: $monospace;
+		font-size: 12px;
+		background-color: #eee;
+		width: 100%;
+	}
+
+	span code {
+		background-color: transparent;
+	}
+}
