+++ client/component/geo-map/index.js
@@ -10,9 +10,10 @@ import classnames from 'classnames';
 /**
  * Internal dependencies
  */
-import Spinner from 'component/wordpress/spinner';
+import Spinner from 'component/spinner';
 import { getMap } from 'state/info/action';
 import { STATUS_IN_PROGRESS, STATUS_FAILED, STATUS_COMPLETE } from 'state/settings/type';
+import './style.scss';
 
 class GeoMap extends React.Component {
 	constructor( props ) {
