+++ client/component/geo-map/style.scss
@@ -1,0 +1,77 @@
+.geo-map {
+	padding-bottom: 10px;
+	width: 100%;
+	position: relative;
+
+	.geo-full {
+		height: 600px;
+	}
+
+	&.geo-map-small {
+		height: 100px;
+		padding-top: 20px;
+	}
+
+	iframe {
+		position: absolute;
+		top: 0;
+		left: 0;
+		width: 100%;
+		background-color: #ddd;
+		height: 450px;
+		max-height: 90%;
+	}
+
+	table {
+		background-color: white;
+		padding: 10px;
+		padding-bottom: 30px;
+		position: absolute;
+		bottom: 0;
+		left: 0;
+		height: 130px;
+		width: 100%;
+	}
+
+	table th,
+	table td {
+		padding: 0;
+	}
+
+	table th {
+		font-weight: bold;
+		text-align: left;
+		width: 150px;
+	}
+
+	table td {
+		text-align: left;
+	}
+
+	h2 {
+		line-height: 1;
+		margin: 0;
+		padding-bottom: 10px;
+	}
+
+	.external {
+		position: absolute;
+		right: 15px;
+		bottom: 10px;
+	}
+}
+
+@media screen and (max-width: 782px) {
+	.modal {
+		.geo-map {
+			iframe {
+				height: 255px;
+			}
+
+			.geo-full {
+				height: 400px !important;
+			}
+		}
+	}
+
+}
