+++ client/component/http-check/style.scss
@@ -1,0 +1,44 @@
+@import '../../global';
+
+.http-check-results {
+	display: flex;
+	padding-bottom: 40px;
+
+	.http-status {
+		width: 80px;
+
+		.dashicons {
+			font-size: 70px;
+			width: 70px;
+			height: 70px;
+		}
+
+		.dashicons-yes {
+			color: $ok;
+		}
+
+		.dashicons-no {
+			color: $error;
+		}
+	}
+
+	.http-result {
+		text-align: left;
+		padding-top: 10px;
+	}
+}
+
+.http-check {
+	h2 {
+		margin-bottom: 0;
+		padding-bottom: 5px;
+		text-align: left;
+		padding-left: 10px;
+	}
+
+	.external {
+		position: absolute;
+		right: 15px;
+		bottom: 15px;
+	}
+}
