+++ client/component/menu/style.scss
@@ -1,0 +1,9 @@
+.subsubsub-container::before,
+.subsubsub-container::after {
+	content: "";
+	display: table;
+}
+
+.subsubsub-container::after {
+	clear: both;
+}
