+++ client/component/spinner/style.scss
