+++ client/component/table/filter.js
@@ -5,7 +5,7 @@
 import React from 'react';
 import { translate as __ } from 'lib/locale';
 import PropTypes from 'prop-types';
-import Select from 'component/wordpress/select';
+import Select from 'component/select';
 
 class TableFilter extends React.Component {
 	static propTypes = {
