+++ client/component/useragent/index.js
@@ -10,9 +10,10 @@ import classnames from 'classnames';
 /**
  * Internal dependencies
  */
-import Spinner from 'component/wordpress/spinner';
+import Spinner from 'component/spinner';
 import { getAgent } from 'state/info/action';
 import { STATUS_IN_PROGRESS, STATUS_FAILED, STATUS_COMPLETE } from 'state/settings/type';
+import './style.scss';
 
 class Useragent extends React.Component {
 	constructor( props ) {
