+++ client/component/useragent/style.scss
@@ -1,0 +1,48 @@
+.useragent {
+	box-sizing: border-box;
+	padding-bottom: 30px;
+
+	th {
+		width: 120px;
+		vertical-align: top;
+		line-height: 1;
+	}
+
+	td {
+		line-height: 1.2;
+	}
+
+	h2 {
+		padding-left: 10px;
+	}
+
+	td,
+	h2 {
+		text-align: left;
+	}
+
+	h2 {
+		margin-bottom: 0;
+		padding-bottom: 5px;
+	}
+
+	.useragent-agent {
+		font-size: 12px;
+	}
+
+	.external {
+		position: absolute;
+		right: 15px;
+		bottom: 15px;
+	}
+
+	.agent-unknown,
+	.agent-unknown h2 {
+		text-align: center;
+		padding: 5px;
+	}
+
+	table {
+		padding-bottom: 15px;
+	}
+}
