+++ client/global.scss
@@ -20,3 +20,4 @@ $green: #4ab866;
 $sans-serif: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen-Sans", "Ubuntu", "Cantarell", "Helvetica Neue", sans-serif;
 $error: #ff3860;
 $ok: #46b450;
+$monospace: courier, Monaco, monospace;
