+++ client/page/support/status.scss
@@ -1,4 +1,4 @@
-@import '../../globals';
+@import '../../global';
 
 /**
  * The plugin status section of the support page
