+++ package.json
@@ -94,6 +94,8 @@
     "sass-loader": "^7.1.0",
     "shared-git-hooks": "^1.2.1",
     "style-loader": "^0.23.1",
+    "stylelint": "^9.7.0",
+    "stylelint-config-wordpress": "^13.1.0",
     "through": "^2.3.8",
     "through2": "^2.0.3",
     "underscore": "^1.9.1",
